+++ neosvr_headless_webui/__init__.py
@@ -4,9 +4,23 @@ from .auth import login_required
 from os import path
 from base64 import b64encode
 
+import logging
+
 def create_app():
     app = Flask(__name__, instance_relative_config=True)
     app.config.from_pyfile("config.py")
+
+    if app.config["LOG_FILE"]:
+        rl = logging.getLogger()
+        rl.setLevel("INFO")
+        hdlr = logging.FileHandler(app.config["LOG_FILE"])
+        fmt = logging.Formatter(
+            fmt="[%(asctime)s] [%(name)s] %(levelname)s: %(message)s",
+            datefmt="%Y-%m-%d %H:%M:%S"
+        )
+        hdlr.setFormatter(fmt)
+        hdlr.addFilter(logging.Filter("neosvr_headless_webui"))
+        rl.addHandler(hdlr)
 
     logo_path = path.join(app.instance_path, app.config["APP_LOGO"])
     if path.exists(logo_path):